Refactored inquiry implementation to better organize code and utilize menu configs and operation flows.
